+++ raiden-dapp/babel.config.js
@@ -8,6 +8,9 @@ if (process.env.E2E) {
 }
 
 module.exports = {
-  presets: [['@vue/cli-plugin-babel/preset', { useBuiltIns: 'entry' }]],
+  presets: [
+    ['@vue/cli-plugin-babel/preset', { useBuiltIns: 'entry' }],
+    '@babel/preset-typescript',
+  ],
   plugins: plugins,
 };