+++ raiden-dapp/CHANGELOG.md
@@ -8,11 +8,13 @@
 
 - [#2178] Backup state reminder notification
 - [#2285] Add custom Cypress attributes to selectors
+- [#2274] Add Cypress tests for closing/settling channel
 
 ### Changed
 
 [#2178]: https://github.com/raiden-network/light-client/issues/2178
 [#2285]: https://github.com/raiden-network/light-client/issues/2285
+[#2274]: https://github.com/raiden-network/light-client/issues/2274
 
 ## [0.12.0] - 2020-10-22